+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 * support windows
 * remove `hii init` step tips
 * update `vue` demo (thanks [zhouhailong](https://github.com/zhouhailong) )
+* support `css` auto prefix
+* support `css`/`less`/`scss` loader config
+* support open browser firefox/safari/opera
+* support launch firefox with pac proxy
+* support launch opera with pac proxy
+
+* fixbug: dll plugin `process.env.NODE_ENV` not right